requirements.txt
pandas
numpy
scikit-learn
pyarrow
pyyaml
lighthgbm
catboost
xgboost
mlflow
matplotlib
seaborn
joblib
pytest
pytest-cov